Python support for object pinning (persistent BPF)
This commit adds support for the BPF_OBJ_PIN system call, already
present in libbpf. A BPF program or map can be made persistent if pinned
to a file system path with the new methods:

  b.pin_func_to_fs(func_name, path)

or

  b.pin_table_to_fs(table_name, path)

Any process with enough permissions will be able to open such path and
gain access to the BPF file descriptor. The BPF object will persist
until the file has been unlinked and all the file descriptors are
closed.

Persistent objects are necessary for iptables and tc BPF commands.

BPF objects can only be pinned to a BPF file system. It can be mounted
with the following command:

  # mount -t bpf none /sys/fs/bpf

In /sys/fs/bpf there is an empty directory for this purpose.

Signed-off-by: Oriol Arcas <oriol@starflownetworks.com>